TimerData * keep only the setRunTime(int) method variant * reworked all calling code

+++ src/main/java/com/xceptance/xlt/api/actions/AbstractAction.java
@@ -361,7 +361,7 @@ public abstract class AbstractAction
                     finally
                     {
                         runTime = TimerUtils.get().getElapsedTime(start);
-                        actionData.setRunTime(runTime);
+                        actionData.setRunTime((int) runTime);
                     }
                 }
                 catch (final Throwable e)
